feat: Add test page for emails
