+++ src/client/js/index.js
@@ -8,5 +8,6 @@ import './resize-observer.js'
 import './nav.js'
 import './user-menu.js'
 import './partials/breaches.js'
+import './partials/emails.js'
 import './partials/settings.js'
 import './components/custom-select.js'
