+++ src/views/main.js
@@ -123,6 +123,7 @@ const mainHeader = data => `
   <a href='/user/breaches' class='nav-item ${data.partial.name === 'breaches' ? 'current' : ''}'>Data breaches</a>
   <a href='/user/data-removal' class='nav-item ${data.partial.name === 'dataRemoval' ? 'current' : ''}'>Data removal</a>
   <a href='/user/settings' class='nav-item ${data.partial.name === 'settings' ? 'current' : ''}'>Settings</a>
+  <a href='/user/emails' class='nav-item ${data.partial.name === 'emails' ? 'current' : ''}'>Emails</a>
 </nav>
 `
